Refactor: 보기쉽게 파일을 분리함
