+++ Kabinett/Presentation/View/WriteLetter/Cells/WriteLetterEnvelopeCell.swift
@@ -72,5 +72,5 @@ struct WriteLetterEnvelopeCell: View {
             )
         }
     }
-
+    
 }
