+++ Kabinett/Presentation/View/WriteLetter/Components/MiniTabBarView.swift
@@ -7,7 +7,7 @@
 
 import SwiftUI
 
-struct MiniTabBar: View {
+struct MiniTabBarView: View {
     @Binding var letterContent: LetterWriteModel
     @ObservedObject var viewModel: ContentWriteViewModel
     @ObservedObject var customTabViewModel: CustomTabViewModel
